# build: upgrade to go 1.23.6