--- a/build/teamcity/internal/release/build-and-publish-patched-go/impl-fips.sh
+++ b/build/teamcity/internal/release/build-and-publish-patched-go/impl-fips.sh
@@ -9,7 +9,7 @@
 set -xeuo pipefail
 
 GO_FIPS_REPO=https://github.com/golang-fips/go
-GO_FIPS_COMMIT=fc8a2bd706bcd6b18b260a9aea59cf63884acdeb
+GO_FIPS_COMMIT=12327118900b0833266189a293cba8ad674901c1
 GOCOMMIT=$(grep -v ^# /bootstrap/commit.txt | head -n1)
 
 # Install build dependencies
@@ -30,14 +30,12 @@ git checkout $GO_FIPS_COMMIT
 # current build infrastructure results in the following error:
 #     version `GLIBC_2.32' not found (required by external/go_sdk_fips/bin/go)
 rm ./patches/017-fix-linkage.patch
+# This patch doesn't apply on Go 1.23.6, but all the files are test-only, so
+# they shouldn't make a difference for the purpose of building the SDK.
+# When golang-fips supports Go 1.23.6, this line can be deleted.
+rm ./patches/023-crypto-tls-fix-config-time.patch
 # Lower the requirements in case we need to bootstrap with an older Go version
 sed -i "s/go mod tidy/go mod tidy -go=1.16/g" scripts/create-secondary-patch.sh
-# We need some extra support so we can clone from our fork rather than the
-# upstream go repo (support for GOLANG_REPO).
-# `golang-fips.patch` contains https://github.com/golang-fips/go/pull/255
-# as a direct patch. This can be removed when GO_FIPS_COMMIT advances beyond
-# 12327118900b0833266189a293cba8ad674901c1.
-git apply /bootstrap/golang-fips.patch
 GOLANG_REPO=https://github.com/cockroachdb/go.git ./scripts/full-initialize-repo.sh "$GOCOMMIT"
 cd go/src
 # add a special version modifier so we can explicitly use it in bazel