--- a/pkg/util/ctxutil/BUILD.bazel
+++ b/pkg/util/ctxutil/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "canceler_1_20.go",
         "canceler_1_21_bazel.go",
         "context.go",
+        "context_linkname.go",
         "doc.go",
         "fast_value.go",
     ],
@@ -21,6 +22,7 @@ go_library(
 go_test(
     name = "ctxutil_test",
     srcs = [
+        "context_linkname_test.go",
         "context_test.go",
         "fast_value_test.go",
     ],