--- a/pkg/util/goschedstats/BUILD.bazel
+++ b/pkg/util/goschedstats/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "goschedstats",
     srcs = [
         "runnable.go",
+        "runnable_common.go",
         "runtime_go1.19.go",
         "runtime_go1.20_21_22.go",
         "runtime_go1.23.go",