… .bazelrc
… DEPS.bzl
… build/bazelutil/distdir_files.bzl
… build/bazelutil/staticcheckanalyzers/def.bzl
… build/bazelutil/staticcheckanalyzers/sa1031/BUILD.bazel
… build/bazelutil/staticcheckanalyzers/sa1031/analyzer.go
… build/bazelutil/staticcheckanalyzers/sa1032/BUILD.bazel
… build/bazelutil/staticcheckanalyzers/sa1032/analyzer.go
… build/bazelutil/staticcheckanalyzers/sa4032/BUILD.bazel
… build/bazelutil/staticcheckanalyzers/sa4032/analyzer.go
… build/bazelutil/staticcheckanalyzers/sa6006/BUILD.bazel
… build/bazelutil/staticcheckanalyzers/sa6006/analyzer.go
… build/bazelutil/staticcheckanalyzers/sa9009/BUILD.bazel
… build/bazelutil/staticcheckanalyzers/sa9009/analyzer.go
… build/teamcity/internal/release/build-and-publish-patched-go/commit.txt
… build/teamcity/internal/release/build-and-publish-patched-go/golang-fips.patch
… go.sum
… pkg/cmd/roachtest/tests/cdc.go
… pkg/kv/kvserver/replica_application_result_test.go
… pkg/rpc/auth_test.go
… pkg/sql/rowexec/processors_test.go
… pkg/util/ctxutil/context.go
… pkg/util/ctxutil/context_linkname_test.go
… pkg/util/ctxutil/context_test.go
… pkg/util/goschedstats/runnable.go
… pkg/util/goschedstats/runnable_test.go
… pkg/util/goschedstats/runtime_go1.23.go
… pkg/util/sysutil/socket.go
… pkg/util/sysutil/socket_stub.go