@dhartunian 2025-02-13
@rickystewart to fix the tracing test please bump the `goid` library version:

`github.com/petermattis/goid v0.0.0-20250211185408-f2b9d978cd7a`

The tracing test relies on identifying the currently running goroutine's id. We do this by replicating the layout of runtime internals (https://github.com/petermattis/goid/blob/master/runtime_go1.23.go to match https://github.com/golang/go/blob/59b7d40774b29bd1da1aa624f13233111aff4ad2/src/runtime/runtime2.go#L458) and reaching in to pluck it out. Occasionally this internal struct changes layout and the `goid` field is not where we expect it to be.