@RaduBerinde 2025-02-15
I ran the sysbench benchmark on this PR against its base (5 runs of 10 minutes each). It looks like about 1% improvement.

Before:
Mean ≈ 23988.90 QPS
Stddev ≈ 237.0 QPS

After:
Mean ≈ 24645.46 QPS
Stddev ≈ 279.2 QPS