@rickystewart 2025-02-24
> I don't get the confusing use of a global. Why not just contextKey{}? I assume a linter complained about it not being a custom type but contextKey{} should do it.

I saw this pattern in the Go stdlib. I assumed this would be better for performance. Doesn't really matter, I switched it to the way you recommend.